utils: explore group/ungroup differences in the comments
Add commentary to note that group and ungroup havesubtly-different
semantics, thanks in part to the loggingof information, and how
much of this logic could be moved to the DataPHA class.

There's also a small re-arrangement of the code, to better make
the case of the similarities and differences between the two.