data: change group flag of associated backgrounds (fix #1657)
This moves the logic for setting the background group/ungroup
status from the UI layer to the DataPHA class. Users can still
change a single dataset by changing the flag directly - i.e.

    pha.grouped = True

rather than

    pha.group()

Of note is the fact that we have to move the log-on-error code
from UI to DataPHA. This is good for the UI code (as the behavior
matches), but is it sensible? We can compare to the notice method
which does something similar (logging potential errors, but not
really ones from the background), so it is not a completely-new
feature to the DataPHA class (that is, logging messages).

Now, the group method needs to worry about whether the background
component has a grouping array: if it doesn't it will raise a
DataErr which we currently convert into a "warning" message
(actually a INFO log). We do not need to do the same for ungroup
as setting grouping to False does not check / care about the
grouping array. We could change this so that instead of using a
try/except block we expicitly check for bkg.grouping is not None,
but I'm not sure it makes much of a difference.