data: setting grouping to None clears the grouped flag (fix #1659)
If the grouping column is set to None then we remove the grouped
flag. It is a corner case.

The implementation requires access to a superclass variable,
which suggests that it should be done another way, but I do not
know how. The problem is that we want to be able to distinguish
calling this during initialization (when we don't want to do
clean-up work) and from general code (when we do). It may be that
we can re-arrange things to avoid this, but I have not the energy
to work on it.