utils: ignore does not note warning twice (fix #1641)
Since ignore() calls notice() then we don't need to override
ignore as well, and can just add the extra logic into notice.
This means we can move the logic from _notice_warnings directly
into the notice call. I took advantage of using a set, as this
is semantically better suited to the check than a list and the
difference in runtime is not important here.

This is a breaking change as it removes a method from the
Session class, but it is explicitly marked as internal (i.e.
starts with _) and is only used in this one place, so I do
not think we need to warn about this change.

The warning message is augmented with the list of the different
analysis settings, as that may prove useful to users in this
presumably-rare situation.