utils: show_bkg could fail (fix #1645)
The problem was that we had routines that did something like

    def foo(idvals, bkg_id):
        for ids in idvals:
            if bkg_id is None:
                bids = [1]
            else:
                bids = [bkg_id]

            for bkg_id in bids:
                oh_no_i_have_overwritten_bkg_id()

So, if you had multiple datasets (ie idvals is not a scalar) then
the input bkg_id setting would be lost after the first dataset
was processed (it's a little more complex than this, but this is
the basis of the bug).

This was found by pylint.

A number of other style/pylint changes were made in these routines,
including:

 - don't call routines we then ignore, so change

        foo = getfoo()
        if arg is not None:
            foo = [arg]

   to a "if arg is None: ... else: ..." style

 - use str(x) rather than x.__str__()

 - use f-strings

 - use idval rather than id when we can (as id is an existing
   symbol in Python, although we are stuck with it as an
   argument to routines)