utils: wrapped models docstring improvements (fix #215)
Previousy help(gauss1d) in Sherpa would not provide anything
helpful. Now it at least mentions the Gauss1D model that is
wrapped, including the first line of the model's docstring,
if set.

I would like to add in a list of the model parameters but
this is not easy to identify (without creating an instance
of the model, which would not be the hardest thing to do but
I have elected not to add this, as it adds complexity, such
as what happens if the instance can not be created?).

It is not clear what will really help the user here,
but this has to be better than the WrappedModel text. See
issue #215 for more information.

There is also a fix for the ModelWrapper class using a
mutable argument in the __init__ call, as mentioned in
issue #95.