CI do not download miniconda
Use the installed version instead. I do not like the need
to constantly re-initialize conda, but we had to do this with
the previous version so it's not adding anything new. I did try
calling 'conda init --all' in an attempt to avoid the need, but
it did not seem to work.