Tests: add testing of regrid/integrate for unop/binop models
What happens if we try to regrid unop/binop models? What about the
integrate setting? This excludes XSPEC models, as these tests are
being added as a separate commit.

Some tests are marked xfail, and some tests will change significantly
if we change the handling of the integrate flag.

There's also some minor code clean up in the tests.