Tests: add XSPEC regrid/integration cases
Most tests actually fail, but are not marked as xfail as I want
to check we know wy they fail (during development of this work
there have been a number of changes to the system, and it's
hard to track what is really happening without explicit tests).

Note that the "combined" test - that is mdl1 * mdl2 - is tested
as both mul * add and add * mul since the regrid code currently
uses the first model when calculating the regrid method which
breaks here. I have also added a test where one of the models
is an ArithmeticConstant model, since they can have different
behavior to "normal" models.

There are also more "esoteric" tests of "deep" expressions,
like (2 * mdl1) / (3 + mdl2) - even if they don't make much
sense (and currently lead to recursion errors), and tests
related to unary and binomial combination of models.

Tests have also been added of combining Sherpa and XSPEC models
(one "additive" and one "multiplicative").

A pytest fixture - xsmodel - has been added to make it a little-bit
easier to create an XSPEC model (select the correct class and then
instatiate it).