XSPEC: rework macros for defining the XSPEC models
A way of indicating a Python function that uses kwargs has been
added to the top-level extension header (KWSEPC). This is then
used in the generic model-extension header, replacing the MODSPEC
definition. The more-interesting change is in the XSPEC extension
header, where again we replace the local MODSPEC definitions

a) re-order them by languange support (FORTRAN then C/C++) rather
   than the somewhat random order before (this is mainly a code
   style issue)

b) remove the table-model specific definitions as they are only
   used in one place (_xspec.cc) and we can just avoid a lot of
   complexity in the macro by hard-coding the logic. This will
   also make removing the pre XSPEC 12.10.1-specific code easier
   when we get to that PR. (this is the more interesting reason)

In a future PR I will update the XSPEC defines from using FCTSPEC
to KWSPEC, but that is left for later.