model: allow simple model deconstruction
Given a model expression like

    m1 * (m2 + m3 * m4)

the model_deconstruct routine will return

    [m * m2, m1 * m3 * m4]

It starts to get tricky for

    m1 / (m2 + m3)
    (m1 + m2) // m3

which we want to leave alone. Note that the aim of this is to
support the "simple" deconstruction case - that is, expand

    a * (b + c + d)

into

    [a * b, a * c, a * d]

and is similar to model expressions that users may write, and
not to be a full expansion/simplification. That is, we do not
want to simplify either of

    a + b - a
    a * (b / a)

to

    b

As the BinaryOpModel class is open - that is, a user could send
in there own version, the code is explicit in only supporting the
numpy binary operators that are supported in the ArithmeticModel
class (e.g. numpy.add, numpy.multiply). Any other binary operator
will not be expanded, even if it is valid.

I have noted that the behaviour is a bit surprising for unary-operator
expressions - that is

   -(a + b * c + d)

is left as is but

   x - (a + b * c + d)

is expanded to

   x, -(a), -(b * c), -(d)

but I am okay with this disconnect as I think that the unary-operator
case is unlikely to be a normal case. We can change the code to support
this but it's likely to complicate it and I'd rather not do so unless
we really have to.

The test checks two things

  1. that the deconstructed terms have the expected form;
     this check is purely syntactic as it checks the name field
     of the model e.g. that the first response is "(b1 * g1)"

  2. a semantic check, by evaluating the model on a grid and
     checking that it matches the summed terms when they are
     evaluated on the same grid

As noted in the test, this is part a functional test - i.e.
does it work - and part a regression test, just to check whether
we get the same thing back. For example, at present there's no
attempt to simplify expressions such as moving the negation
operator or trying to combine contant terms.

A possibility is to add property testing (as popularised in the
python community by the hypothesis package), where random
expressions would be created and then the semantic check ran
on them (i.e. that the model evaluation matches the sum of
the terms), but the hope is that the existing test cases are
sufficient.

There is currently no use if this functionality, but it is
planned for new plot_model/source_components calls.