utils: minor code clean up
Avoid using a fully-qualified name when we have the symbol
imported anyway.

Clean up the use of _clean_model (make more use of it) and add a
little documentation.

Explicitly import sherpa.models.model as we require it. This comes
from sherpa.all but I find this hard to follow so would prefer to
make things explicit.