Tests: re-work how show output is tested
The existing code is annoying to change whenever a new field is
added to the middle of the output. Rework these tests to make it
a bit-more declarative. There's a choice over how complex we want
to make this, so I've decided not to move the getvalue call into
the check_line* calls, to make it easier to use in different
scenarios.