fake: rework fake_pha
Address issues #1682 (documentation) and #1685 (behavior). The
functional changes are primarily to sherpa.astro.fake.fake_pha,
where the API has been significantly changed, but there has been
a less-extensive change to the API of sherpa.astro.utils.fake_pha.

Note that the sherpa.astro.fake.fake_pha routine is "for experts", so
a large API change is, I believe, acceptable. The old arguments are
still present - so no code will fail because of the changes here - but
user's should now see a warning message, and the code may behave
differently. I am still not particularly happy with the API, but I
feel it is better, and I am not sure it's worth improving (although
that is in part because the complexity of the UI fake_pha routine has
increased).

The sherpa.astro.ui.utils.fake_pha routine is the main user API, so
the change here is less extensive, but it does involve a new argument
(the method argument, which users are not expected to change, and was
actually added in #1734), and - because bugs have been fixed - the
output may be different. I am also not happy about the API of this
routine, as it should be possible to take more data from the existing
PHA object (e.g. response and exposure time), and some thought is
needed on how we really want to handle backrounds (e.g. how we include
the background contribution in the source, how do we handle multiple
backgrounds, and do we want to easily allow the background datasets to
be simulated by this routine).

The significant changes to sherpa.astro.fake.fake_pha are

- the model must now include any response term, which removes the
  need for the is_source and pileup_model arguments
- if background models are to be included then it's up to the
  caller to include them in the model parameter
- in order to allow the simulation of the background from the
  background counts (which is likely to be noisy, but is an option)
  I had thought about only doing it if the PHA object contained
  background datasets, but have decided that that might make
  upstream code more annoying to deal with, so have gone with the
  include_bkg_data flag. Given the mess that is
  sherpa.astro.ui.utils.fake_pha I may decide that we do not
  need include_bkg_data.