Tests: extend the fake_pha checks
Check that we error out. The main idea is to check the "PHA has no
response" error check, but we also add two regression tests that check
what happens if the first two arguments are not a PHA or model,
respectively. These are only regression tests as we do not have an
explicit check, so the error is caught "indirectly". The checks are
explicit about that they are looking for so we will know if there is
any later change.

There are also tests that are added or amended to test out the
handling of backgrounds, both for the case where the data is
sampled and for when background models are used. Note that two tests
are now known to test incorrect behavior, but they are left as
passing in this, to act as a regression test. They will be updated
in a later commit.