Tests: add checks based on #1685 and #1686
Add fake_pha tests which explicitly check the handling of the
bgnd argument. There are other tests that do something similar,
but I have not tried to consolidate them.

Add an explicit check on what get_source returns when a background
model is set. This is related to #1686 and is just a check of the
current behavior. It is also relevant to some of the changes
planned in #1684.