Tests: add checks related to #1688
The issue behind #1688 is that the #1672 change had a line like

    warning("foo %s bar", val)

and this is being displayed as "foo %s bar". So, test_set_hard_limits
was written to test this. However, it turns out that the logging output
that iscaught by caplog does include the string interpolation, so this
check (while useful), is not going to show that #1688 is fixed.

When writing this Iinitially got confused and ended up adding
test_set_pars_errors__soft_limits as a check, which is actually
urelated to #1688 but I've left it in.