sherpa: rework default logger formatter
After

  >>> from sherpa.models.basic import Box1D
  >>> mdl = Box1D('m')

before this commit we had

  >>> mdl.thawedpars = [-4e40, 4e40, -4e40]
  WARNING: value of parameter %s is below minimum; setting to minimum
  WARNING: value of parameter %s is above maximum; setting to maximum
  WARNING: value of parameter %s is below minimum; setting to minimum

but we now have

  >>> mdl.thawedpars = [-4e40, 4e40, -4e40]
  WARNING: value of parameter m.xlow is below minimum; setting to minimum
  WARNING: value of parameter m.xhi is above maximum; setting to maximum
  WARNING: value of parameter m.ampl is below minimum; setting to minimum

The default logger formatter, which is used to add the level to the
message (if above INFO level), dropped all-but the first argument to a
log call. This has been okay, but in #1672 calls were changed to use
the lazy-formatting form of

    warning("blah %s blah", "blah")

form, which ended up being output as

    blah %s blah

Changing the formatter to using getMessage() to retrieve the message
appears to fix the issue.

An alternative would be to always ensure we use a single-argument call
to the logging routines, but there's no way to easily/sensibly enforce
this.

I have tried my best to add a test for this, but I have not found
a way to intercept the stdout from the logger:

- using caplog just captures the logging output created by the default
  formatter, which doesn't suffer from this problem and hence doesn't
  show the issue

- using capsys or capfd does not capture the output, and I do not
  understand why (since some part of pytest can catch it)

I am going to rely on this being tested by external-to-Sherpa testing,
which is how #1688 was found anyway (on-going testing of CIAO
development code by SDS).