utils: fake_pha API additions
The fake_pha command can now be called with just an id value,
with the other values taken from the original PHA dataset. The
main changes are

- the arf, rmf, and exposure arguments are now given default
  values
- an exposure argument of None means that the DataPHA exposure
  value is used; this is a change, as previously it would set
  the DataPHA exposure value to None.

We allow the exposure value to be None, which is not particularly
useful but is how it used to be (I had at one point made it error
out for this case but decided against it).