image: control the DS9 instance name
Connecting to DS9 requires a "name" - identified as the template value
in the sherpa.image.DS9 interface. This name is used to identify the
DS9 via XPA.

The current interface is to use the value of the
sherpa.image.DS9._DefTemplate variable, which has been set to
"sherpa". This means that any Sherpa session on your machine will talk
to the *same* DS9 instance. This is one way to transfer information
between the two sessions, but it is not the best way, and is more
likely to cause problems. However, we haven't had anyone complain
about it yet.

This commit adds get_template() and set_template() calls that will
return the template name, or change the template name *and* update the
current imager to use the new backend name. The old window is *NOT
CLOSED DOWN*. which is not what we really want.

This is one way to allow the Sherpa tests to be run in parallel, but
it is not a guaranteed solution (since it depends on how the
multiprocessing code splits up the tests into separate pid sessions).

There is also some test changes:

The test_ds9 test is a bit weird, as it is not clear what it's really
testing compared to the test_image routine. It has to create a DS9
instance, so we update the test to use the new get_template routine
for setting the template name (it actually does not matter what the
name is, but I wanted to make it a bit-more like how we create the DS9
instance).