image: control the DS9 instance name
Connecting to DS9 requires a "name" - identified as the template value
in the sherpa.image.DS9 interface. This name is used to identify the
DS9 via XPA.

The current interface is to use the value of the
sherpa.image.DS9._DefTemplate variable, which has been set to
"sherpa". This means that any Sherpa session on your machine will talk
to the *same* DS9 instance. This is one way to transfer information
between the two sessions, but it is not the best way, and is more
likely to cause problems. However, we haven't had anyone complain
about it yet.

This commit adds get_template() and set_template() calls that will
return the template name, or change the template name *and* update the
current imager to use the new backend name. The old window is *NOT
CLOSED DOWN*. which  may not be what we really want, but at least the
user can close the instance if they do not want it without deleting
it if the user actually wants to keep the old instance open.

Note that there is no attempt (at present) to make this set_template
routine available direclty from the UI code. If the user wants it
then they will have to import it directly from thesh erpa.image module
or the ds9_backend.

This is one way to allow the Sherpa tests to be run in parallel, but
it is not a guaranteed solution (since it depends on how the
multiprocessing code splits up the tests into separate pid sessions).