data: cleanup of background id handling
This is a minor clean up, mainly

- calling the variable bkg_id or resp_id rather than id where we
  can do this without changing the API (i.e. only within a
  function)
- avoid copying of the background/responise_ids array, changing
  the copy, and them overwriting with this new copy when we can
  just operate in place
- explicit convert an iterable to a list so that when it gets
  converted to a string in an error message it reads nicely

I had intended to simplify the handling of the bkg_ids values,
as we do not need to keep around a separate list but could just
use the keys of the background dict. However, that would stop one
feature we currently support (and have added a test for) - that
is, I can load multiple backgounds, then change background_ids
to remove a subset, then add them back in without having to
re-create the backgrounds. I am not sure how useful this is, but
I've decided to leave as is. See the tests

- test_pha_check_background_ids_basic
- test_pha_check_respnose_ids_basic

in sherpa/astro/tests/test_astro_data2.py