plot: add sherpa_plot context manager
Creating a sherpa plot follows the same template, and it fits in
well to being a context manager (even though there is no "plot
object" we have to attach the open/close logic to, we can use the
contextmanager decorator).

This is mainly useful internally, but it may be useful outside of
Sherpa.