UTILS: allow plots to accept multiple ids WIP
Argh: the code explicitly copies the plot structues, so that
we can safely modify them after-the-fact. And because we need this
if we have multiple ids, otherwise we just end up with the same
plot repeated n times.

Unfortunately this breaks the get_xxx_plot ethos, as it isn't
going to match the actually-plotted version.

Although I want to keep recalc=False I also want it to go away.
Far away...

What about the astro versions? In partcular the background variants?

What about contours?