plot: move plot code out of Session class (internal)
This is an internal change. The plotting code does not need to
be a member of the session class, so move it out. It could have
been marked as a classmethod, but we have enough methods there
anyway.

Rather than remove the _plot and _contour methods they have been
marked deprecated.