ensure a background is a DataPHA object
By explicily checkingthat the background is a DataPHA object we

- provide a nicer user experience
- don't leave the object in a partially-converted state

This does not implement all possible checks (e.g. that the source
and background regions have the same set of channels) as I haven't
yet thought through the implications.

The dataspace1d docstring has been reworked as the last example,
which created DataPHA objects, was a bit confused and tried to
set the channels to an energy range. This example has been re-worked
to use channels instead.