data: ensure background and source data use the same channels
Technically the background could have a different set of channels,
particularly as the user can filter them, but it does not seem
worth supporting this complexity. As we currently implicitly
assume that the sets of channels agree (e.g. when subtract is
called) let's just make it a requirement.

This had a surprisinggly large change for the testing, mainly
to remove or clean up tests that are now less relevant (or in fact
are no-longer relevant).