Tests: add tests showing #1721
So, get_stat_info misses data and returns invalid data when we have
a combination of PHA with background and PHA with no background. So
add a test of this. This is a regression test, even though we know
it's wrong, so we can see when it gets fixed.

There are also a bunch of related tests that check the fit output.
None of these seem to be affected directly by #1721 but there is
one case where the output is sub-optimal (lists all the ids when
it really should only list the id with the background model).