utils: get_bkg_model will use source dataset for response
The get_bkg_model call needs to add a response in (for the case
when set_bkg_model and not set_bkg_full_model was used). It used
to only use the background PHA, assuming that the code had set
things up beforehand so that the background had a respose. This
is normally true, as the code takes some pain to assign a
respnose to the background from the source data, but this is
not infallible - e.g. adding responses after adding the background.

So, we grab the response from the source data if it isn't present
in the background. I note that _add_extra_data_and_models - which
is used by fit and fit_bkg - does this, so we should do it here
as well.