Tests: internal change to how logging is handled
The test_astro_ui.py test turned off logging for all tests, and
then selectively turned on the logging when required (under the
assumption that we got no log messages at the ERROR level).

However, this means that we don't get to test the logging in
general, only for those calls we care about, and this is a bit
awkward here. Note that these are "ui-level" tests, where it is
helpful to note when a log message has changed, added, or been
removed. With the 4.15.0 release (and planned 4.15.1) we have more
logging here than nbefore (e.g. the notice/ignore and group family
of calls), so it makes sense to

a) remove the hide_logging fixture that was used for all the tests
   in this file
b) move to using SherpaVerbosity rather than caplog.at_level
   for determining the log level we want (it does the same thing
   but we should use the Sherpa functionality)
c) abstract the testing - e.g. add the exact_record and match_record
   calls - to make it a bit cleaner

There is a question of whether we should check each log message.
I've decided no - we can skip some when we know they have been
tested elsewhere.