plot: should the user be told about a missing plot backend? WIP
The code now warns user's if the requested plot backend can not
be loaded (as long as it's not "dummy") with a line like

  >>> from sherpa import ui
  WARNING: Failed to import any usable plot backend: tried 'pylab'

There is the question that if the plot_pkg is "pylab dummy" then
should it not warn here, since the user has explicitly said "I am
okay with there being no plot backend"?

There is the follow-on question: should the PLOT_BACKENDS list
include "dummy" as a synonym for "BasicBackend"?