Tests: add a backend plot test for a basic PHA plot
This replicates the sherpa.astro.ui.plot_fit call for a simple
PHA dataset, but using the OO interface. It was added as the
test_plot_model_all_backends test from
sherpa/astro/ui/tests/test_astro_ui_plot.py was failing, and
this helps determine whether it's a UI issue or not.