utils: easy way to change plot backend at the ui layer WIP
The idea is to make it easy to change the plot backend for a UI
user. There's two parts

- simplify things for the user
- add a place where the code can determine what to do

A user can always change the backend manually (with the set_backend
call from sherpa.plot) but then they are responsible for any
downstream damage.

A version of the all_plot_backends fixture is created for the
astro ui case - all_plot_backends_astro_ui - which uses the
set_plot_backend call - but no test needs to use it. Which may
mean we don't need all the machinery here, but maybe we do?