Tests: enhance xspec user-model tests
Check handling of several cases that were broken (or poorly
handled) in the 4.15 release

- case sensitivity of FORTRAN model names
- what happens when the model is marked
  - dataset dependent
  - requires the errors

The problem was that the tests were too permissive, in an attempt
to simplify them. It turned out to be a bad idea, so the tests
now just check the expected output as a whole, rather than trying
to cherry-pick the things they were interested in.