Tests: add a check for #1240
Issue #1240 was strange, and appears to have been an issue in
XSPEC rather than Sherpa, but add a check to see if it happens
again.

The test code is also updated to using the tmp_path fixture from
pytest rather than the tempfile module, as it is a bit nicer /
more hygenic.