utils: use dictionary rather than a tuple
Provide some level of structure to the data sent around when
determing what data and models to fit by switching from a tuple
to a dict. This doesn't significantly improve the code - in fact
it makes it slightly longer - but it is more extensible: with a
tuple all we really have to go on is the length, but with a dict
we can check what keys are present (not that we actually need
this extra functionality, but it makes reviewing the code easier).