utils: use a dataclass rather than a dictionary
This extends the idea of providing structure to the data sent
around when working out what to fit by using dataclasses rather
than dictionaries. This makes the code slightly-shorter, but I
think it's more-important at expressing intent.

The sherpa.ui.utils.Session._get_fit_ids call is a bit of an
anomaly here as as we use a dictionary. It could create a FitStore
with no model, but that removes the point of saying "this object
contains all we need to fit this dataset".