utils: re-work fitting internals
I had changed _get_fit_ids to return the data as well as the
identifier, but it doesn't really lead to much of a code saving,
so I removed. I did think about removing _get_fit_ids entirely
but it's just-about complicated enough that I don't want to have
to replicate the logic in both _prepare_fit and _prepare_bkg_fit.