Tests: add a check of fit vs fit_bkg differences
When reviewing the code I noted that the handling of background
datasets missing a model component were handler differently by fit
and fit_bkg. This adds two tests to explicitly test this (although
the functionality may well be tested elsewhere).

There are arguments they should behace the same or differently,
so this is just meant to be a canary to check the existing behavior.