utils: fix #1722
The problem was that resp_ids=range(1) which ends up being the same
as resp_ids=[0]. Unfortunately when there's only one response we
assume that it has the default id (which is 1), so a call like

    pha.get_response()

will then fail with a message about there being no response. The
better fix is to ensure we pass the resp_id argument to the
get_response call, but that is a *significantly* larger job, as
there's a lot of them. So this is a fix for just #1722 (which
should still work if we address the get_response issue later).