ui: add set_rng/get_rng methods
Allow the session to contain a generator that is used to create
random numbers. It is set with the set_rng call and can be
retrived with get_rng.

If set_rng(None) is called - which is the default for sessions -
then the existing legaxy NumPy random API is used.

For users who set the random seed with

    np.random.seed(seed)

then they can get the same results with

    set_rng(np.random.RandomState(seed))

but it's suggested that you use either of

    set_rng(np.random.default_rng(seed))
    set_rng(np.random.default_rng())

depending on whether you want something that is repeatable (altough
note that the generator does not guarantee that a given seed will
always generate the same results in the future).

Note that thanks to these changes the reset_seed fixture has been
removed.