Add rng parameter to most random-number-using routines
In most cases we can send in a rng parameter which, if set to None
(the default), will use the legacy NumPy random API - e.g. call
np.random.xxx directly. If set, then rng is assumed to be a
generator, and used to access the random scheme directly.

Issues are:

a) this is not used by the "higher-up" routines yet
b) the main area not addressed here is sherpa.optmethods

A number of routines have been added to sherpa.utils.random to
help in this transition (that is, when the rng parameter may be
None). Thses are intended to be "short-term" features as they
can be replaced with dirct calls to the object.

Note that code that ends up calling the new parallel_map_rng
routine will create different results when sent a RandomState
object (to re-create the legacy results) as the processes are now
each sent an explicit and separate generator compared to the old
system where they were all using the implicit seed.