sim: ensure that most arguments to get_sample are keyword only
The ParameterSample class defines get_sample with a signature
accepting fit, num=1. This is an abstract class and could be
marked as such, but for now it is left as is.

Derived classes have a variety of arguments, and unfortunately
they do not always extend the fit, num=1 pattern. For example,
UniformParameterSampleFromScaleVector uses the arguments fit,
factor=4, num=1. This means that any call to get_sample has
to be careful about the arguments it sends (other than the fit
argument, which we assume is always the first argument).

To make the code easier to read, and to avoid worrying about
possible issues - such as "is it 'dof, num' or 'num, dof'" -
this commit

a) explicitly labels all arguments to get_sample other than
   the first argument

   This leads to a very-minor code increase as there's at least
   one place where I have had to replace a generic

       samples = sampler.get_sample(fit, arg, num)

   since the second argument (i.e. arg) is named mycov in one
   case and myscales in the other case.

b) marked all-but-the first argument as keyword-only - that is,
   callers must pass in the keyword.

The test_sim.py file has seen a minor clean up (there was a
try/except block that was only needed for debugging, and would
actually re-throw the error it was trying to catch half way
through, so I removed it).