utils: move parallel code into sherpa.utils.parallel
The sherpa.utils code is rather diverse, so this commit moves out
the "parallel" code into sherpa.utils.parallel, and then imports
it so it's still visible from sherpa.utils (although the idea would
be to change the calling code to use the new location).

There are some small style changes to the code (mainly the use of
f-strings and replace a lambda with an actual function, which
does not need to be send the data to process as it's already in
scope). There is one change to avoid sending a Lock object around
as we end up not using it (to match the recent change in the
optimiser code).

Move some of the logic from run_tasks into a separate routine,
process_tasks - which is not in the export list, as this code can
be used by the specialized parallel code in estmethods.

The num argument for run_tasks turns out to not be needed (it is
always the same as the length of the procs argument) so this has
been made an optional argument, with the note that it will be
deprecated at some point in the future (but no time limit given,
and not marked as an actual deprecated value).

There's some attempt to make the methods - such as the internal
my_worker helper routine in the sherpa/opt code - to be "better
typed" by having more-explicit arguments. More could be done, but
I'm not sure it's worth it here.

One interesting part of this is that the multi-processing setup has
been moved out of sherpa/utils/__init__.py. In #975 I suggested
moving this code into a separate module. For this commit I decided
not to do this, to make it easier to trace the behavior.

This appears to have fixed #1744 - at least the second part. To
be honest I am not sure why: I would have expected we still need
to check for possible errors even after processing the success
queue, and the specifics of #1744 part 2 are mysterious to me, so
I do not know what could have changed in this commit to "fix"
things. I think the fix is that we no longer use the num argument
to determine the length of the output structure (that is because
the run_tasks routine now uses other information to get this value,
but this is just a guess).