optmethods: internal change to be explicit about values being sent
This should not change behavior, as it just ensures the constructors
for the SimplexBase-related classes are called with most of the
arguments benig explicitly named - e.g. Foo(a=1, b=2, c=3) rather
than Foo(1, 2, 3). This is to make it easier to understand what is
going on with the seed value (to make sure that values are sent
around).

The SimplexBase class has an init method which is the main routine
which is over-ridden in derived classes. These classes all send in a
factor argument so add it to the base routine (to please any
lint-style analysis).

There are some minor code-style changes to make the code a little
more readable (e.g. no need to send a parameter through to temporary
routines when we know it's value already, and minor loop changes to
be more Pythonic.