opt: pass random numbers with the rng option
As the code now uses the NumPy RNG code, the seed value used in
the sherpa/optmethods/tests/test_opt_unit.py will create different
random values. The test has been updated to show that when using
np.random.RandomState we get the same value as with the seed
setting. A new test has been added to show the "new" RNG interface.
This will need updating when they change the code (so it may end
up being more effort than it's worth).

Note that some code passes the seed around and it is not used
with the np.random.seed call:

- to control the random numbers from the random module
- sent to C++ code for the MT random code

We do not want to (or, rather, are not going to) change the second
option. For these we just need to ensure we send in a seed (and,
if necessary, have this generated randomly).

The first option can be changed. This means that any code using a
particular seed will no-longer be repeatable, as we are changing
from random.seed to numpy.random.seed (when not sending in an
actual generator).

This is hard to check and verify because there is limited
coverage of this code and it's not obvious what is going on, or
whether the seed/generator is being sent around. There's also
at least two places that looks broken that have been noted but
not changed here

- a seed argument is sent in but it is then replaced by a fixed
  value
- a call is missing a required seed argument