pyfits: be more lenient with FITS files with messed-up headers
It is easy to end up with a FITS file with an invalid header but
valid data. Astropy does allow the header to be "fixed up" when
certain problems are found, so do so.

There is a possibility that this verification should be done at a
different location, but without attempting to rework the module
I elected for the "verify at the point we care about the header"
approach.

The other choice is whether any fixes should be silent or let the
user know there was something funky: I elected for silentfix as the
screen output when there is a problem is rather verbose.