XSPEC: report parameter values when an XSPEC model fails
This should not happen, but occasionally an XSPEC model will fail.
In this case it is useful to know what the model was and what the
parameters were. As an example (this is with XSPEC 12.13.0 and will
hopefully be addressed in an XSPEC patch or update, maybe just by
bumping up the minimum value of the tmax parameter [*]):

[*] this was originally written before the bug was understood, but
    after my diagnose to the HEASARC team they have adjusted the tmax
    minimum so it doesn't cause the problem.

    >>> from sherpa.astro import xspec
    >>> mdl = xspec.XScevmkl()
    >>> mdl.tmax = 0.01

- before

    >>> mdl([0.1, 0.2, 0.3], [0.3, 0.4, 0.5])
    Traceback (most recent call last):
      ...
    ValueError: XSPEC model evaluation failed

- with this commit

    >>> mdl([0.1, 0.2, 0.3], [0.3, 0.4, 0.5])
    Traceback (most recent call last):
      ...
    ValueError: XSPEC additive model evaluation failed: xscevmkl.cevmkl alpha=1.0 Tmax=0.01 nH=1.0 He=1.0 C=1.0 N=1.0 O=1.0 Ne=1.0 Na=1.0 Mg=1.0 Al=1.0 Si=1.0 S=1.0 Ar=1.0 Ca=1.0 Fe=1.0 Ni=1.0 redshift=0.0 switch=1.0 norm=1.0

Note that

a) it now tells you the "type" of the model (additive, multiplicative,
   convolution, or table)
b) the class and model names are included
c) the parameters are listed, along with their parameter names
   (this can make it hard to cut-n-paste into code, but makes it
   a lot easier to see that it's Tmax that has a strange value,
   for example)

The original version of the code attempted to make the text appear
in the C++ code. This turns out to be tricky as the C++ code has
no idea what the Python model is going to be called. There is also
a difference in how table model evaluations work at the C++ layer.
So I ended up adding it to the Python layer, as we have all the
information we need. It does mean that users who call the model
directly - e.g. call sherpa.astro.xspec._xspec.funcname directly -
will not benefit from this, but they are expected to be able to
handle this case themselves, as it is an "internal" API.

I also tried to add a sub-exception of ValueError, so we knew
when to add the message, but it doesn't seem to really help and
there were a lot of problems in passing the exception object
around in the C++ code (i.e. it would have been declared and
created in _xspec.cc but used in xspec_extension.hh).

This is hard to test (you need a model you know has a problem,
which the XSPEC models should not have) so I am not 100% sure it'll
work in all cases.