Make doctestplus a tests_require dependency
Add `pytest-doctestplus` to the requirements for tests.

Effectively, this will make all our doctests run by default.

We merged lazy loading of modules #1631 which allows us to run doctestplus
(if installed) locally and test that the examples in docstrings and sphinx docs
work, which at the same time uses them as separate tests.
Since that works, this PR explores making doctestplus a testing requirement,
which would install it on also on every CI run and thus
get the doctests checked in CO as well.