parallel: add parallel_map_rng call
Allow processes that take a rng parameter to be called in parallel
with a separate generator per call. The code has not been updated
to use this routine yet.

Note that the tests for this method are very limited and do not
actually test that a separate generator is sent per process.