fit: code clean up
Primarily changing code to use f-strings, but also

- create a list of strings that then get concataneted via
  "\n".join(...) rather than repeated `s += ....` lines
- skip a pass case and only deal with the error condition
- take advantage of "lazy logging" - e.g. use
  warning("foo %s %s", b1, b2) - rather than a f-string (this
  is mainly to make pylint happy); also use implicit string
  concatenation - that is "foo " "bar" is the same as
  "foo " + "bar"

There are some places where the "old" string formtting code is
retained, as it is just easier to keep this than re-write it.

pylint points out that two classes have NoNewAttributesAfterInit as
the superclass but do not call it's initalizer. I have added a note
about this but have not tried to make this change. There's also
an import within a function that could probably be top level, but
I'm converned about possible cyclic imports so have not checked
(but have added a note).