fit: add rng argument - is this meaningful
At the moment I haven't tracked down where the fit actually happens
as there's the switch from fit to iterfit which calls back to fit.

So I don't see where the rng needs to get passed through. It is
also at this point we need to switch from rng to seed values, as
the optimisers use a seed (for the C++ code at least).