Tests: check the simulation results
The test_sim.py test was just running code and not checking the
results. So we now

- explicitly set the seed,so it should be repeatable
  (hopefully across platforms)
- check results

Many of the checks are regression tests - that is, they will need
updating if things like the random generator changes. The results
look okay, but as I did not write these tests in the first place
I can not guarantee this. There are some useful tests - e.g. that
multivariate_t returns a subset of the data returned by t_sample
(this holds for a number of routines).

The test_lrt test has been split into two: one where numcores=1
and one with numcores=2, test_lrt_multicore (which will end up
running the same code as numcores=1 when multiprocessing is not
supported, such as on our CI runs). The latter test is more a
"does the code just return" rather than a full test, as we can not
guaranteee the RNG used with the current code. However, there
are still a number of issues

- #1746 suggests that the model values should be reset after
  each iteration, which should help avoid getting stuck in some
  part of the search space
- the model evaluation is sensitive to the platform (macOS vs
  linux) which means that the fit can end up with different
  parameter values, which can lead to differences in the ratio
  calculated for an iteration (not clear how this is related to
  issue #1746).